requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
crewai==0.5.0
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langsmith==0.0.87
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.5.0
mypy_extensions==1.1.0
narwhals==1.43.0
numpy==1.26.4
openai==1.88.0
packaging==23.2
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.46.0
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
yarl==1.20.1
